\name{plot.target}
\alias{plot.target}
\title{Plot a zero-targeted model +/- 1 standard deviation.}
\usage{
  plot.target(x, raw = F, ...)
}
\arguments{
  \item{x}{a zero-targeted model object}

  \item{raw}{boolean indicating whether to plot raw values;
  defaults to FALSE (i.e. print scaled rather than raw
  values)}
}
\description{
  Plot a zero-targeted model +/- 1 standard deviation.
}

