\name{unpack_formula}
\alias{unpack_formula}
\title{Unpack an S3 formula to figure out the name of the DV and IVs}
\usage{
  unpack_formula(formula)
}
\arguments{
  \item{formula}{model specification}
}
\value{
  target object
}
\description{
  Unpack an S3 formula to figure out the name of the DV and
  IVs
}

